the end and other bug storries

+++ views/admin/courses/index.php
@@ -26,8 +26,8 @@
                     <td><?php echo $course->department_name; ?></td>
                     <td><span class="badge bg-success"><?php echo $course->status; ?></span></td>
                     <td>
-                        <a href="<?php echo BASE_URL; ?>admin/courses/edit/<?php echo $course->id; ?>" class="btn btn-sm btn-warning"><i class="fas fa-edit"></i></a>
-                        <a href="<?php echo BASE_URL; ?>admin/courses/delete/<?php echo $course->id; ?>" class="btn btn-sm btn-danger"><i class="fas fa-trash"></i></a>
+                        <a href="<?php echo BASE_URL; ?>admin/courses/edit/<?php echo $course->id; ?>" class="btn btn-sm btn-warning" data-bs-toggle="tooltip" title="Edit"><i class="fas fa-edit"></i></a>
+                        <a href="<?php echo BASE_URL; ?>admin/courses/delete/<?php echo $course->id; ?>" class="btn btn-sm btn-danger" data-bs-toggle="tooltip" title="Delete" data-confirm="Are you sure you want to delete this course?"><i class="fas fa-trash"></i></a>
                     </td>
                 </tr>
                 <?php endforeach; ?>
