+++ views/admin/users/index.php
@@ -34,8 +34,8 @@
                     <td><?php echo $user->department_id; ?></td>
                     <td><span class="badge bg-success"><?php echo $user->status; ?></span></td>
                     <td>
-                        <a href="<?php echo BASE_URL; ?>admin/users/edit/<?php echo $user->id; ?>" class="btn btn-sm btn-warning"><i class="fas fa-edit"></i></a>
-                        <a href="<?php echo BASE_URL; ?>admin/users/delete/<?php echo $user->id; ?>" class="btn btn-sm btn-danger"><i class="fas fa-trash"></i></a>
+                        <a href="<?php echo BASE_URL; ?>admin/users/edit/<?php echo $user->id; ?>" class="btn btn-sm btn-warning" data-bs-toggle="tooltip" title="Edit"><i class="fas fa-edit"></i></a>
+                        <a href="<?php echo BASE_URL; ?>admin/users/delete/<?php echo $user->id; ?>" class="btn btn-sm btn-danger" data-bs-toggle="tooltip" title="Delete" data-confirm="Are you sure you want to delete this user?"><i class="fas fa-trash"></i></a>
                     </td>
                 </tr>
                 <?php endforeach; ?>
